e-learning-platform-main/e-learning-platform-main/src/main/java/de/buw/se/backend/model/User.java
package de.buw.se.backend.model;

public class User {

    private int id;
    private String email;
    private String username;
    private String fullName;
    private String password;

    // Constructor
    public User(int id, String email, String username, String fullName, String password) {
        this.id = id;
        this.email = email;
        this.username = username;
        this.fullName = fullName;
        this.password = password;
    }

    // Getters and Setters
    public int getId() {
        return id;
    }

    public void setId(int id) {
        this.id = id;
    }

    public String getEmail() {
        return email;
    }

    public void setEmail(String email) {
        this.email = email;
    }

    public String getUsername() {
        return username;
    }

    public String getFullName() {
        return fullName;
    }

    public void setFullName(String fullName) {
        this.fullName = fullName;
    }

    public void setUsername(String username) {
        this.username = username;
    }

    public String getPassword() {
        return password;
    }

    public void setPassword(String password) {
        this.password = password;
    }
}
